+++ src/commands/config.rs
@@ -1,5 +1,5 @@
 use crate::commands::arg_structs::config_args::{BoolArgs, CommonArgs, ConfigCommand, ConfigSubCommand};
-use crate::config_manager::{get_config};
+use crate::config_manager::{get_config, Config};
 use crate::utils::{command_output, display_table, get_expanded_path, CellData};
 use std::path::PathBuf;
 use tracing::{warn};
@@ -10,17 +10,13 @@ pub fn parse_config_args(config_cmd: &ConfigCommand) {
             set(&args.common)
         },
         ConfigSubCommand::List => {
-            println!("listing all configurations");
-        },
-        ConfigSubCommand::Show(_args) => {
-            // show(&args.common);
+            list()
         },
         ConfigSubCommand::Del(args) => {
-            println!("{:?}", args);
+            del(args)
         },
     }
 }
-
 
 fn set(args: &CommonArgs) {
 
@@ -40,11 +36,11 @@ fn set(args: &CommonArgs) {
         }
     }
 
-    if let Some(show) = &args.notify_of_unzipped_mods {
-        config.notify_of_unzipped_mods = *show;
+    if let Some(notif) = &args.notify_of_unzipped_mods {
+        config.notify_of_unzipped_mods = *notif;
         config.save(None).unwrap();
 
-       display_vec.push(command_output("config.show_mod_dir_warning".to_string(), show.to_string()));
+       display_vec.push(command_output("config.show_mod_dir_warning".to_string(), notif.to_string()));
     }
 
     if let Some(version) = &args.pin_game_version {
@@ -93,38 +89,60 @@ fn set(args: &CommonArgs) {
 
 }
 
-// fn show(args: &CommonArgs) {
-//
-//         if let Some(mod_dir) = {}
-//         CommonArgs { pin_game_version: String, ..} => {}
-//         CommonArgs { zip_mod_dirs: bool, .. } => {}
-//         CommonArgs { backup_mods: bool, .. } => {}
-//         CommonArgs { backup_mods_dir: String, ..} => {}
-// }
 
-fn _list() {
-    println!("listing all configurations...");
-}
+fn del(args: &BoolArgs) {
 
-fn _del(args: &BoolArgs) {
+    let mut config = get_config().write().unwrap();
+    let defaults = Config::default();
+    let mut display_vec: Vec<(CellData, CellData)> = Vec::new();
+
+
     if args.backup_mods_dir {
-
+        config.backup_mods_dir = defaults.backup_mods_dir.clone();
+        display_vec.push(command_output("config.backup_mods_dir".to_string(), defaults.backup_mods_dir.to_string()));
     }
 
     if args.zip_mod_dirs {
-
+        config.zip_mod_files = defaults.zip_mod_files.clone();
+        display_vec.push(command_output("config.zip_mod_files".to_string(), defaults.zip_mod_files.to_string()));
     }
 
     if args.backup_mods {
-
+        config.backup_mods = defaults.backup_mods.clone();
+        display_vec.push(command_output("config.backup_mods".to_string(), defaults.backup_mods.to_string()));
     }
 
     if args.pin_game_version {
-
+        config.pinned_game_version = defaults.pinned_game_version.clone();
+        display_vec.push(command_output("config.pinned_game_version".to_string(), defaults.pinned_game_version.to_string()));
     }
 
-    if args.mods_dir {
-
+    if args.mod_dir {
+        config.mod_dir = defaults.mod_dir.clone();
+        display_vec.push(command_output("config.mod_dir".to_string(), defaults.mod_dir.to_string()));
     }
 
+    if args.notify_of_unzipped_mods {
+        config.notify_of_unzipped_mods = defaults.notify_of_unzipped_mods.clone();
+        display_vec.push(command_output("config.notify_of_unzipped_mods".to_string(), config.notify_of_unzipped_mods.to_string()));
+    }
+
+
+    display_table(display_vec, None);
+    config.save(None).unwrap();
+}
+
+fn list() {
+    let config = get_config().read().unwrap();
+    let mut display_vec: Vec<(CellData, CellData)> = Vec::new();
+
+    display_vec.push(command_output("config.mod_dir".to_string(), config.mod_dir.to_string()));
+    display_vec.push(command_output("config.backup_mods".to_string(), config.backup_mods.to_string()));
+    display_vec.push(command_output("config.backup_mods_dir".to_string(), config.backup_mods_dir.to_string()));
+    display_vec.push(command_output("config.zip_mod_files".to_string(), config.zip_mod_files.to_string()));
+    display_vec.push(command_output("config.show_execution_time".to_string(), config.show_execution_time.to_string()));
+    display_vec.push(command_output("config.notify_of_unzipped_mods".to_string(), config.notify_of_unzipped_mods.to_string()));
+    display_vec.push(command_output("config.pinned_game_version".to_string(), config.pinned_game_version.to_string()));
+
+    display_table(display_vec, None);
 }